requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
openpyxl>=3.1.2
python-docx>=1.1.0
pymupdf>=1.23.0
google-genai>=0.3.0
pydantic>=2.6.0
python-dotenv>=1.0.0
